refactor: rename agro_robot_sim → devkit_simulation and anglicize launch names
- Rename package agro_robot_sim → devkit_simulation
- Update package.xml, CMakeLists.txt, launch files, config, and worlds
- Update all references in devkit_launch/sowbot_sim.launch.py, devkit_ui/ui_node.py,
  manage.py, README.md, and docker/Dockerfile
- Anglicize launch file names:
  missao_fazenda.py → farm_mission.py
  gravador_rotas.py → route_recorder.py
  fazenda_completa.launch.py → complete_farm.launch.py
