--- src/agro_robot_sim/ament_flake8.ini
@@ -1,7 +1,0 @@
-[flake8]
-extend-ignore = B902,C816,CNL100,D100,D101,D102,D103,D104,D105,D106,D107,D203,D212,D404,I101,I202,Q000
-import-order-style = google
-max-line-length = 99
-show-source = true
-statistics = true
-jobs = 1
