--- src/agro_robot_sim/config/gps_ekf.yaml
@@ -1,54 +1,0 @@
-### gps_ekf.yaml ###
-navsat_transform:
-  ros__parameters:
-    frequency: 30.0
-    magnetic_declination_radians: 0.0  # Ajuste se for vida real. Simulação é 0.
-    yaw_offset: 0.0                    # Aponta para o norte na simulação
-    zero_altitude: true
-    broadcast_cartesian_transform: true
-    publish_filtered_gps: true
-    use_odometry_yaw: false
-    wait_for_datum: false
-    
-    # Frames
-    base_link_frame: base_link
-    world_frame: odom      # O navsat converte GPS para odometria global
-
-ekf_filter_node_map:
-  ros__parameters:
-    frequency: 30.0
-    two_d_mode: true
-    publish_tf: true
-    
-    map_frame: map
-    odom_frame: odom
-    base_link_frame: base_link
-    world_frame: map
-
-    # 1. Odometria das rodas (Diff Drive)
-    odom0: /odom
-    odom0_config: [false, false, false,
-                   false, false, false,
-                   true,  true,  false,
-                   false, false, true,
-                   false, false, false]
-    odom0_differential: false
-
-    # 2. IMU (Giroscópio/Acelerômetro)
-    imu0: /imu
-    imu0_config: [false, false, false,
-                  true,  true,  true,
-                  false, false, false,
-                  true,  true,  true,
-                  true,  true,  true]
-    imu0_differential: false
-    imu0_remove_gravitational_acceleration: true
-
-    # 3. GPS (Vindo do navsat_transform_node já convertido em metros)
-    odom1: /odometry/gps
-    odom1_config: [true,  true,  false,
-                   false, false, false,
-                   false, false, false,
-                   false, false, false,
-                   false, false, false]
-    odom1_differential: false
