--- src/agro_robot_sim/config/ros_gz_bridge.yaml
@@ -1,65 +1,0 @@
-# ros_gz_bridge configuration for agro_robot_sim (Gazebo Harmonic / ROS 2 Jazzy)
-#
-# parameter_bridge argument strings do NOT support gz topic remapping via the
-# "gz.msgs.Type:/model/name/topic" colon-suffix notation — that syntax is parsed
-# as part of the type name and produces "No template specialization for the pair".
-# Use this YAML config instead; load it with:
-#   ros2 run ros_gz_bridge parameter_bridge --ros-args -p config_file:=<path>
-#
-# NOTE: the model's plugins (DiffDrive, JointStatePublisher, sensors) publish on
-# UNSCOPED gz topics (/cmd_vel, /joint_states, /odom, /imu, /gps, /scan), NOT
-# /model/agro_robot/*. Verified with `gz topic -i`. Names must match exactly or
-# the bridge connects to nothing.
-
-- ros_topic_name: /cmd_vel
-  gz_topic_name: /cmd_vel
-  ros_type_name: geometry_msgs/msg/Twist
-  gz_type_name: gz.msgs.Twist
-  direction: ROS_TO_GZ
-
-- ros_topic_name: /odom
-  gz_topic_name: /odom
-  ros_type_name: nav_msgs/msg/Odometry
-  gz_type_name: gz.msgs.Odometry
-  direction: GZ_TO_ROS
-
-# DiffDrive has <tf_topic>/tf</tf_topic> (absolute) so it stays on the global /tf.
-- ros_topic_name: /tf
-  gz_topic_name: /tf
-  ros_type_name: tf2_msgs/msg/TFMessage
-  gz_type_name: gz.msgs.Pose_V
-  direction: GZ_TO_ROS
-  publisher_qos:
-    reliability: reliable
-
-- ros_topic_name: /scan
-  gz_topic_name: /scan
-  ros_type_name: sensor_msgs/msg/LaserScan
-  gz_type_name: gz.msgs.LaserScan
-  direction: GZ_TO_ROS
-
-- ros_topic_name: /imu
-  gz_topic_name: /imu
-  ros_type_name: sensor_msgs/msg/Imu
-  gz_type_name: gz.msgs.IMU
-  direction: GZ_TO_ROS
-
-- ros_topic_name: /gps/fix
-  gz_topic_name: /gps
-  ros_type_name: sensor_msgs/msg/NavSatFix
-  gz_type_name: gz.msgs.NavSat
-  direction: GZ_TO_ROS
-
-- ros_topic_name: /joint_states
-  gz_topic_name: /joint_states
-  ros_type_name: sensor_msgs/msg/JointState
-  gz_type_name: gz.msgs.Model
-  direction: GZ_TO_ROS
-
-- ros_topic_name: /clock
-  gz_topic_name: /clock
-  ros_type_name: rosgraph_msgs/msg/Clock
-  gz_type_name: gz.msgs.Clock
-  direction: GZ_TO_ROS
-  publisher_qos:
-    reliability: reliable
