--- src/agro_robot_sim/worlds/empty.sdf
@@ -1,56 +1,0 @@
-<?xml version="1.0" ?>
-<!--
-  empty world for Gazebo Harmonic with sensor plugins.
-  > gz-sim-sensors-system
-  > gz-sim-rendering-system
-  > gz-sim-scene-broadcaster-system
--->
-<sdf version="1.8">
-  <world name="empty">
-
-    <plugin filename="gz-sim-physics-system"
-            name="gz::sim::systems::Physics"/>
-
-    <plugin filename="gz-sim-user-commands-system"
-            name="gz::sim::systems::UserCommands"/>
-
-    <!-- gz-sim-sensors-system -->
-    <plugin filename="gz-sim-sensors-system"
-            name="gz::sim::systems::Sensors">
-      <render_engine>ogre2</render_engine>
-    </plugin>
-
-    <!-- gz-sim-rendering-system -->
-    <plugin filename="gz-sim-rendering-system"
-            name="gz::sim::systems::Rendering"/>
-
-    <!-- gz-sim-scene-broadcaster-system -->
-    <plugin filename="gz-sim-scene-broadcaster-system"
-            name="gz::sim::systems::SceneBroadcaster"/>
-
-    <light name="sun" type="directional">
-      <cast_shadows>true</cast_shadows>
-      <pose>0 0 10 0 0 0</pose>
-      <diffuse>0.8 0.8 0.8 1</diffuse>
-      <specular>0.2 0.2 0.2 1</specular>
-      <direction>-0.5 0.1 -0.9</direction>
-    </light>
-
-    <model name="ground_plane">
-      <static>true</static>
-      <link name="link">
-        <collision name="collision">
-          <geometry><plane><normal>0 0 1</normal></plane></geometry>
-        </collision>
-        <visual name="visual">
-          <geometry><plane><normal>0 0 1</normal><size>100 100</size></plane></geometry>
-          <material>
-            <ambient>0.8 0.8 0.8 1</ambient>
-            <diffuse>0.8 0.8 0.8 1</diffuse>
-          </material>
-        </visual>
-      </link>
-    </model>
-
-  </world>
-</sdf>
