+++ src/devkit_simulation/package.xml
@@ -1,0 +1,28 @@
+<?xml version="1.0"?>
+<?xml-model href="http://download.ros.org/schema/package_format3.xsd" schematypens="http://www.w3.org/2001/xml-model"?>
+<package format="3">
+  <name>devkit_simulation</name>
+  <version>0.2.0</version>
+  <description>Simulation package for the Feldfreund DevKit — ROS 2 Jazzy + Gazebo Harmonic</description>
+  <maintainer email="joaodemouragy@gmail.com">joaodemouragy-hash</maintainer>
+  <license>Apache-2.0</license>
+
+  <buildtool_depend>ament_cmake</buildtool_depend>
+
+  <depend>robot_state_publisher</depend>
+  <depend>xacro</depend>
+  <depend>joint_state_publisher</depend>
+  <depend>geometry_msgs</depend>
+  <depend>sensor_msgs</depend>
+
+  <!-- Gazebo Harmonic (replaces gazebo_ros / gazebo_ros_pkgs) -->
+  <depend>ros_gz_sim</depend>
+  <depend>ros_gz_bridge</depend>
+
+  <test_depend>ament_lint_auto</test_depend>
+  <test_depend>ament_lint_common</test_depend>
+
+  <export>
+    <build_type>ament_cmake</build_type>
+  </export>
+</package>
